requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==2.3.2
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
cssselect2==0.7.0
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-db-connection-pool==1.2.5
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==5.4.0
django-silk==5.3.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==35.2.0
flower==2.0.1
fonttools==4.56.0
gprof2dot==2025.4.14
greenlet==3.2.1
gunicorn==23.0.0
html5lib==1.1
humanize==4.11.0
idna==3.10
kombu==5.4.2
lxml==5.3.1
numpy==2.2.3
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycodestyle==2.13.0
pycountry==24.6.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pypdf==5.3.1
pyphen==0.17.2
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
sqlparams==6.2.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
xlrd==2.0.1
XlsxWriter==3.2.2
zopfli==0.2.3.post1
